+++ crates/ds3-archipelago/src/core.rs
@@ -51,6 +51,7 @@ pub struct Core {
 
 impl shared::Core for Core {
     type SlotData = SlotData;
+    type Game = crate::game::DarkSoulsIII;
 
     /// Creates a new instance of the mod.
     fn new() -> Result<Self> {